Create routes for authentication and separate server file.
