+++ Projectwork/package.json
@@ -15,6 +15,7 @@
     "dotenv": "^16.3.1",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
+    "express-flash": "^0.0.2",
     "express-session": "^1.17.3",
     "mongoose": "^7.6.3",
     "passport": "^0.6.0",
